% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\name{get_image}
\alias{get_image}
\title{Image Crawler.}
\usage{
get_image(address)
}
\arguments{
\item{address}{A webpage address.}
}
\value{
A list of image address.
}
\description{
Crawl all the image address as a list of string from the webpage.
}
\details{
Input webpage address, return all the image address inside the webpage.
}
\examples{
get_image("https://en.wanweibaike.com/wiki-University\%20of\%20Toronto")
get_image("https://en.wanweibaike.com/wiki-Toronto")
get_image("https://en.wanweibaike.com/wiki-list\%20of\%20largest\%20cities")
}
