% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\name{get_table}
\alias{get_table}
\title{Table Crawler.}
\usage{
get_table(address)
}
\arguments{
\item{address}{A webpage address.}
}
\value{
A list of dataframe.
}
\description{
Crawl all the tables as dataframe from the webpage.
}
\details{
Input webpage address, return all the table as a list inside the webpage.
}
\examples{
get_table("https://en.wanweibaike.com/wiki-List\%20of\%20ambassadors\%20to\%20Canada")
get_table("https://en.wanweibaike.com/wiki-List_of_House_members_of_the_42nd_Parliament_of_Canada")
get_table("https://en.wanweibaike.com/wiki-List\%20of\%20largest\%20cities")
}
