% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\name{get_text}
\alias{get_text}
\title{Text Crawler.}
\usage{
get_text(address)
}
\arguments{
\item{address}{A webpage address.}
}
\value{
A string of text.
}
\description{
Crawl all the text as a string from the webpage.
}
\details{
Input webpage address, return all the text inside the webpage.
}
\examples{
get_text("https://en.wanweibaike.com/wiki-Canada.")
get_text("https://en.wanweibaike.com/wiki-Pacific_Ocean")
get_text("https://en.wanweibaike.com/wiki-Parliament_of_the_United_Kingdom")
}
